% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ambdata}
\alias{Ambdata}
\title{Ambiguity-Conflict data}
\format{A data frame with 166 rows and 2 variables:
\describe{
  \item{ID}{subject ID}
  \item{value}{Rating in each judgment scenario}
  \item{scenario}{Index for judgment scenarios}
}}
\source{
\url{http://psycnet.apa.org/record/2006-03820-004}
}
\usage{
Ambdata
}
\description{
A data from a study that investigates the judgment under ambiguity and conflict
}
\keyword{datasets}
