% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrLRT.R
\name{anova.cdfqr}
\alias{anova.cdfqr}
\title{Model comparison test for fitted cdfqr models}
\usage{
\method{anova}{cdfqr}(object, ..., test = "LRT")
}
\arguments{
\item{object}{The fitted cdfqr model.}

\item{...}{One or more cdfqr model objects for model comparison.}

\item{test}{The model comparison test, currently only 'LRT' is implemented.}
}
\description{
Likelihood Ratio Tests for fitted cdfqr Objects.
}
\examples{
data(cdfqrExampleData)
fit_null <- cdfquantreg(crc99 ~ 1 | 1, 't2','t2', data = JurorData)
fit_mod1 <- cdfquantreg(crc99 ~ vert | confl, 't2','t2', data = JurorData)
anova(fit_null, fit_mod1)

}
