% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugsLikelihood.R
\name{bugsLikelihood}
\alias{bugsLikelihood}
\title{Likelihood Functions for Generating OpenBUGS Model File}
\usage{
bugsLikelihood(fd, sd)
}
\arguments{
\item{fd}{A string that specifies the parent distribution.}

\item{sd}{A string that specifies the sub-family distribution.}
}
\value{
A string to be written in the BUGS model file.
}
\description{
Likelihood functions for generating OpenBUGS model file.
}
\examples{
bugsLikelihood('t2','t2')
}
