% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPCC}
\alias{IPCC}
\title{IPCC data-set}
\format{A data frame with 4014 rows and 8 variables:
\describe{
  \item{subj}{Subject ID number}
  \item{treat}{Experimental conditions}
  \item{valence}{Valence of the sentences}
  \item{prob}{raw probability estimates} 
  \item{probm}{Linear transformed prob into (0, 1) interval} 
  \item{mid}{Distinguish lower, best and upper estiamtes }
  \item{high}{Distinguish lower, best and upper estiamtes } 
  \item{Question}{IPCC question number} 
}}
\source{
\url{http://journals.sagepub.com/doi/abs/10.1111/j.1467-9280.2009.02284.x}
}
\usage{
IPCC
}
\description{
The IPCC data-set comprises the lower, best, and upper estimates 
for the phrases "likely" and "unlikely" in six IPCC report sentences.
}
\keyword{datasets}
