% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPCCAUS}
\alias{IPCCAUS}
\title{IPCC data-set - Australian data}
\format{
A data frame with 4014 rows and 8 variables:
\describe{
  \item{ID}{Subject ID}
  \item{gender}{Gender of subjects, `0`is male, `1`is female}
  \item{age}{age of subjects}
  \item{cfprob}{personal probability.}
  \item{bestprob}{nominated probability.}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/19207697/}
}
\usage{
IPCCAUS
}
\description{
The IPCC-AUS data-set  comprises the best estimates for the phrases 
in IPCC report sentences.
}
\keyword{datasets}
