% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrFitted.R
\name{predict.cdfqr}
\alias{fitted.cdfqr}
\alias{predict.cdfqr}
\title{Methods for Cdfqr Objects}
\usage{
\method{predict}{cdfqr}(object, type = c("full", "mu", "sigma"), ...)

\method{fitted}{cdfqr}(object, type = c("full", "mu", "sigma"), ...)
}
\arguments{
\item{object}{A cdfqr model fit object}

\item{type}{A character that indicates whether the full model prediction/fitted values are needed, or values for the `mu` and `sigma` submodel only.}

\item{...}{currently ignored}
}
\description{
Methods for obtaining the fitted/predicted values for a fitted cdfqr object.
}
\examples{
data(cdfqrExampleData)
fit <- cdfquantreg(crc99 ~ vert | confl, 't2','t2', data = JurorData)

plot(predict(fit))

}

