% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrLogLik.R
\name{qrLogLik}
\alias{qrLogLik}
\title{Log Likelihood for Fitting Cdfquantile Distributions}
\usage{
qrLogLik(y, mu, sigma, fd, sd)
}
\arguments{
\item{y}{the vector to be evaluated.}

\item{mu}{mean of the distribution.}

\item{sigma}{sigma of the distribution.}

\item{fd}{A string that specifies the parent distribution.}

\item{sd}{A string that specifies the sub-family distribution.}
}
\value{
The negative log likelihood for fitting the data with a cdfquantile distribution.
}
\description{
Function to give the (negative) log likelihood for fitting cdfquantile distributions.
}
\examples{
y <- rbeta(20, 0.5, 0.5)
qrLogLik(y, mu = 0.5, sigma = 1, 't2','t2')

}

