% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yoon}
\alias{yoon}
\title{Patient Time Data}
\format{
A data frame with 1170 rows and 11 variables:
\describe{
  \item{id}{case identification}
  \item{Day}{day of the week ( 0 = Sunday)}
  \item{Ambulance}{0 = walk-in; 1 = ambulance-arrival}
  \item{Triage}{triage level}   
  \item{Triage1}{1 = triage level 1}  
  \item{Triage2}{1 = triage level 2}  
  \item{Triage3}{1 = triage level 3}  
  \item{Triage4}{1 = triage level 4}  
  \item{Triage5}{1 = triage level 5}  
  \item{Lab}{1 = laboratory test(s) conducted}  
  \item{Xray}{1 = x-ray conducted}
  \item{Other}{1 = other intervention}
  \item{LOS}{length of stay in minutes}
  \item{LOSh}{length of stay in hours}
  \item{preg}{proportion of time in registration stage}
  \item{ptriage}{proportion of time in triage stage}
  \item{pnurse}{proportion of time in nursing care stage}
  \item{pphysician}{proportion of time in consultation with physician(s)}
  \item{pdecis}{proportion of time in decisional stage}
  \item{pregptriage}{preg + ptriage}
  \item{pphysdecis}{pphysician + pdecis}
  \item{prnurse}{pnurse/(pnurse + pregptriage)}
  \item{prphysdec}{pphysdecis /(pphysdecis + pregptriage)}
}
}
\source{
\doi{10.1017/S1481803500006539}
}
\usage{
yoon
}
\description{
Data from Modeling Proportion of Patient Time in Emergency Ward Stages
}
\keyword{datasets}
