% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AnxStrData}
\alias{AnxStrData}
\title{Stress-Anxiety data}
\format{A data frame with 166 rows and 2 variables:
\describe{
  \item{Anxiety}{Scores on Anxiety subscale}
  \item{Stress}{Scores on Stress subscale}
}}
\source{
\url{https://dl.dropboxusercontent.com/u/1857674/betareg/betareg.html}
}
\usage{
AnxStrData
}
\description{
A second example is a data from a study that investigates the relationship 
between stress and anxiety..
}
\keyword{datasets}

